Exceptions mechanism (#539)
Create new class Error and several derived classes: RuntimeError, ValueError, TypeError, etc. * * * These new classes allow stream-like interface to build nice error messages.
* Use new class everywhere in the code. (This should eliminate most of the warnings on Linux builds)
* New macro MAKE_EXCEPTION_SAFE to replace CATCH_EXCEPTIONS. For now this new macro is only used in few places.
* Fixed a memory leak in PyColumn class (although it is unlikely to have impacted anything).

+++ c/fread.c
@@ -1006,7 +1006,7 @@ int freadMain(freadMainArgs _args)
     }
   }
   if (verbose) {
-    if (args.skipNrow) DTPRINT("  skip lines = %lld\n", args.skipNrow);
+    if (args.skipNrow) DTPRINT("  skip lines = %lld\n", (long long)args.skipNrow);
     if (args.skipString) DTPRINT("  skip to string = \"%s\"\n", args.skipString);
     DTPRINT("  showProgress = %d\n", args.showProgress);
   }
